Lots of changes reading the description
Made and implemented settings. Made login in screen and register but the
register is not yet implemented. The app now starts with Login Screen
and then if the 'sign in' or 'continue without signing in' is clicked
then language select screen is brought up. when flag is clicked the main
menu is brought up.